변경된 테이블 구조에 맞게 리팩토링

+++ src/main/java/com/view/zib/domain/address/domain/Coordinate.java
@@ -2,9 +2,11 @@ package com.view.zib.domain.address.domain;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 public record Coordinate(
-        double latitude,
-        double longitude
+        BigDecimal latitude,
+        BigDecimal longitude
 ) {
 
     @Builder
