+++ src/main/java/com/view/zib/domain/api/kako/domain/Coordinate.java
@@ -1,4 +1,6 @@
 package com.view.zib.domain.api.kako.domain;
 
-public record Coordinate(double latitude, double longitude) {
+import java.math.BigDecimal;
+
+public record Coordinate(BigDecimal latitude, BigDecimal longitude) {
 }
